Refactoring command.go
